Move AlarmModal.txt pages -> components
